// packages/sdk/src/demo/data/lot-photos.generated.ts
export const lotPhotos: Readonly<Record<number, readonly MediaImage[]>> = {
  "7601": [
    {
      "src": "/images/lots/7601-brisane-21/01.webp",
      "width": 2000,
      "height": 1333,
      "focalPoint": "40% 55%",
      "alt": {
        "fr": "Petit croiseur sur sa remorque",
        "en": "Small cruiser on its trailer"
      }
    },
// ...
};